move OrgList component to its own file, extract BaseOrg type
