+++ frontend/src/components/pages/profile/ProfileCard/variants/OrgCard/types.ts
@@ -1,0 +1,3 @@
+import { Organization } from "@interfaces/organizations";
+
+export type BaseOrg = Pick<Organization, "id" | "name" | "logoUrl">;
